Moved all foundryvtt config out of declarations and into module
