+++ src/declarations/cosmere-rpg/data/combatant/index.d.ts
@@ -3,8 +3,3 @@ export declare const config: {
     base: typeof CombatantDataModel;
 };
 export { CombatantDataModel } from './combatant';
-declare module "@league-of-foundry-developers/foundry-vtt-types/configuration" {
-    interface DataModelConfig {
-        Combatant: typeof config;
-    }
-}
