+++ src/declarations/cosmere-rpg/data/item/index.d.ts
@@ -46,24 +46,3 @@ export * from './trait';
 export * from './goal';
 export * from './power';
 export * from './talent-tree';
-declare module "@league-of-foundry-developers/foundry-vtt-types/configuration" {
-    interface DataModelConfig {
-        Item: {
-            [ItemType.Weapon]: typeof WeaponItemDataModel;
-            [ItemType.Armor]: typeof ArmorItemDataModel;
-            [ItemType.Equipment]: typeof EquipmentItemDataModel;
-            [ItemType.Loot]: typeof LootItemDataModel;
-            [ItemType.Ancestry]: typeof AncestryItemDataModel;
-            [ItemType.Culture]: typeof CultureItemDataModel;
-            [ItemType.Path]: typeof PathItemDataModel;
-            [ItemType.Talent]: typeof TalentItemDataModel;
-            [ItemType.Trait]: typeof TraitItemDataModel;
-            [ItemType.Action]: typeof ActionItemDataModel;
-            [ItemType.Injury]: typeof InjuryItemDataModel;
-            [ItemType.Connection]: typeof ConnectionItemDataModel;
-            [ItemType.Goal]: typeof GoalItemDataModel;
-            [ItemType.Power]: typeof PowerItemDataModel;
-            [ItemType.TalentTree]: typeof TalentTreeItemDataModel;
-        };
-    }
-}
