+++ src/declarations/cosmere-rpg/documents/actor.d.ts
@@ -170,35 +170,3 @@ export declare class CosmereActor<T extends CommonActorDataModel = CommonActorDa
      */
     private migrateGoals;
 }
-declare module '@league-of-foundry-developers/foundry-vtt-types/configuration' {
-    interface ConfiguredActor<SubType extends Actor.SubType> {
-        document: CosmereActor;
-    }
-
-    interface FlagConfig {
-        Actor: {
-            'cosmere-rpg': {
-                sheet: object;
-                'sheet.mode': 'edit' | 'view';
-                'sheet.expertisesCollapsed': boolean;
-                'sheet.immunitiesCollapsed': boolean;
-                'sheet.skillsCollapsed': boolean;
-                'sheet.hideUnranked': boolean;
-                'sheet.autosetPrototypeTokenValues': boolean;
-                goals: object;
-                'goals.hide-completed': boolean;
-                [key: `meta.update.mode.${string}`]: string;
-                [key: `mode.${string}`]: string;
-            };
-        };
-
-        TableResult: {
-            [SYSTEM_ID]: {
-                'injury-data': {
-                    type: InjuryType;
-                    durationFormula: string;
-                };
-            };
-        };
-    }
-}
