+++ src/declarations/cosmere-rpg/documents/item.d.ts
@@ -311,26 +311,3 @@ type WeaponItem = CosmereItem<WeaponItemData>;
 type GoalItem = CosmereItem<GoalItemData>;
 type PowerItem = CosmereItem<PowerItemData>;
 type TalentTreeItem = CosmereItem<TalentTreeItemData>;
-
-declare module '@league-of-foundry-developers/foundry-vtt-types/configuration' {
-    interface ConfiguredItem<SubType extends Item.SubType> {
-        document: CosmereItem;
-    }
-
-    interface FlagConfig {
-        Item: {
-            [SYSTEM_ID]: {
-                sheet: {
-                    mode: 'edit' | 'view';
-                };
-                'sheet.mode': 'edit' | 'view';
-                meta: {
-                    origin: ItemOrigin;
-                };
-                'meta.origin': ItemOrigin;
-                previousLevel?: number;
-                isStartingPath?: boolean;
-            };
-        };
-    }
-}
