+++ src/declarations/cosmere-rpg/types/hooks/enricher.d.ts
@@ -9,12 +9,3 @@ import { HOOKS } from '@system/constants/hooks';
 type TriggerEnricher = (actorId: string, source: string, data: Record<string, string>) => void;
 export type TriggerTestEnricher = TriggerEnricher;
 export type TriggerDamageEnricher = TriggerEnricher;
-declare module "@league-of-foundry-developers/foundry-vtt-types/configuration" {
-    namespace Hooks {
-        interface HookConfig {
-            [HOOKS.TRIGGER_TEST_ENRICHER]: TriggerTestEnricher;
-            [HOOKS.TRIGGER_DAMAGE_ENRICHER]: TriggerDamageEnricher;
-        }
-    }
-}
-export {};
